requirements.txt
langchain
python-dotenv
langchain-openai
langchain-community
langchain-core
pypdf
pymupdf
faiss-cpu
langchain-groq
langgraph
streamlit